+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * When displaying cozy-home from Cozy's native application, the Support Us is not displayed
 * Upgrade Cozy-Scripts to enable service-worker
+* Photos: Fix pagination issue
 
 ## 🐛 Bug Fixes